refactored unittest import logic to improve accuracy of the coverage measurement
The unittest module is selectively imported from the standard library
if we're running with a Python >= 3.2 or from and external library
otherwise.  This logic was moved from distutils2.test.support to
distutils2.tests because the former module also imports various
modules from distutils2, which would skew the coverage measurement
because the definitions in those module would run before the
instrumentation had a chance to be started.