Adding methods to generate a distutils setup.py file reading the setup.cfg.
That way, distribution maintainers will be able to offer distutils & distutils2
compatibility, without having to maintain the same information in two files.

The only that needs to be maintained is the setup.cfg. Indeed, the setup.py
reads the setup.cfg file every time it is called, to determine which are the
correct args to pass on to distutils.core.setup().