Change resource files declaration in setup.cfg
Resource files declaration move from they own section ``[resources]`` to a
``resources`` key in the ``[files]`` section. The format of the declaration in
the ``resources`` key stay the same::

    [<base-glob>] <glob> = {dispatch}/destination/

this change is motivated by the fact that config parser "key" are always
lower case and can't hold file path information propertly.