Fix backport changesets part 3: backported modules.
shutil, sysconfig, tarfile and their tests have been updated to the
latest 3.2 version (except for test_tarfile which is not backported yet)
and edited to be compatible with 2.4.

Duplicates added in util during the Great Update have been deleted, as
well as functions I removed recently in packaging.  Unneeded modules in
_backport have been deleted or moved to d2.compat.