Refactor support code for compiling xxmodule.c.
This file is needed in other tests, so it’s better to have the support
code in tests.support.  It’s also simpler to just have a skip instead of
custom print/return/test suite fiddling.  Unfortunately, the xxmodule.c
file (resurrected from the repo, and also identical to the version in
Python 2.7) cannot be compiled by Python 2.4 and 2.5 on my computer, so
the test is skipped.

The code to fix up build_ext for Unix shared builds and Windows debug
builds was also moved to support for future reuse.

Finally, I fixed code using sysconfig._CONFIG_VARS directly so that it
calls get_config_var first, so that _CONFIG_VARS is a dict instead of
None.