Update test to account for recent Python bugfix.
This does not mean that we can run tests successfully with a Python 2.7
built from its repository: The version that will become 2.7.3 reports
itself as 2.7.2 (except for a “+” in sys.version, but that’s an
implementation detail).  Some tests need access to Python.h and that
isn’t found when running from an uninstalled Python anyway.