Improve byte-compilation to be independent of -O or -B.
All code (util.byte_compile, build_py, install_lib) can now create .pyc
and/or.pyo files according to options given by users, without
interference from the calling Python’s own optimize mode or from the
sys.dont_write_bytecode switch.

The rationale is that packaging gives control over the creation of
.pyc/.pyo files to the user with its own explicit option, and the
behavior should not be changed if the calling Python happens to run with
-B or -O for whatever reason.

This is actually a bug fix, not an improvement: Digging into the early
history of distutils shows that the original author wanted this behavior
(see for example comments in build_py in r12940).