Move copies of stdlib test code from tests.__init__ to runtests.
I deleted captured_stdout because its usage felt clumsy to me; better to
use stdlib’s test.support.captured_stdout in packaging and just write
out the code for d2.

I checked that adding the import of d2.tests.unittest at the top level
did not change the coverage accuracy (see comment in the source and the
history of the file for more info).