Make test_manifest pass on 2.4 and 2.5 (fixes #11751).
The code added massages the output of fnmatch in old Pythons to make it
match more recent behavior (it was changed in 2.6 to append '\Z(?ms)'
instead of '$', to support newline characters in filenames).  I could
have changed the tests, but I prefer having the code do the same thing
for all versions as far as possible.