Use sys.version_info instead of sys.version.
The contents of this attribute are an implementation detail, as
documented for #9442, so we should not parse it, to support non-CPython
VMs in the future.

Unfortunately, one use comes directly from PEP 345, so an edit will have
to be agreed before fixing the code (see comment in d2.markers).

Other remaining uses are found in d2.compiler and could be replaced by
the platform module (which also parses sys.version, but then it wouldn’t
be my fault :)