Remove out-of-date documentation.
Tarek and I agreed that maintaining doc in two places was wasteful, so
I’m going to point people to the packaging docs in Python 3.3.
Most of them apply as is to distutils2, except for the module name and a
few other things that I’ll fix later (like the logger name or the HTTP
user agent string).

I will put up a simple page with links on packages.python.org to replace
the old doc.